refactor: extract ReviewWorkspaceTimelineEntry.swift from ReviewWorkspaceTypes.swift
Splits the `ReviewWorkspaceTimelineEntry` struct out of
ReviewWorkspaceTypes.swift; that file now holds only the
ReviewWorkspaceTab enum + the ReviewSummaryIssueGroup value.

Byte-identical move of lines 25-38. Both files import Foundation.

ReviewWorkspaceTypes.swift: 73 → 58 lines. Tests: 667.
