+++ Sources/BugNarrator/Utilities/ReviewWorkspaceTimelineEntry.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+struct ReviewWorkspaceTimelineEntry: Identifiable, Equatable {
+    let id: UUID
+    let timestamp: TimeInterval
+    let kind: ReviewWorkspaceTimelineEntryKind
+    let title: String?
+    let text: String
+    let secondaryText: String?
+    let index: Int?
+    let screenshotID: UUID?
+
+    var timeLabel: String {
+        ElapsedTimeFormatter.string(from: timestamp)
+    }
+}
