+++ Sources/BugNarrator/Utilities/ReviewWorkspaceTypes.swift
@@ -22,21 +22,6 @@ enum ReviewWorkspaceTab: Identifiable, CaseIterable {
     }
 }
 
-struct ReviewWorkspaceTimelineEntry: Identifiable, Equatable {
-    let id: UUID
-    let timestamp: TimeInterval
-    let kind: ReviewWorkspaceTimelineEntryKind
-    let title: String?
-    let text: String
-    let secondaryText: String?
-    let index: Int?
-    let screenshotID: UUID?
-
-    var timeLabel: String {
-        ElapsedTimeFormatter.string(from: timestamp)
-    }
-}
-
 enum ReviewWorkspaceTimelineEntryKind: Equatable {
     case transcript
     case marker
